+++ src/components/header-nav/header-nav.tsx
@@ -2,7 +2,7 @@ import { useAppSelector } from '../../hooks';
 import User from './components/user/user';
 import Login from './components/login/login';
 import { AuthorizationStatus } from '../../const';
-import { selectAuthorizationStatus } from '../../store/user-process/selectors';
+import { selectAuthorizationStatus } from '../../store/user-process/user-process.selectors';
 
 function HeaderNav(): JSX.Element{
   const authorizationStatus = useAppSelector(selectAuthorizationStatus);
