+++ src/components/places-sort-form/places-sort-form.tsx
@@ -1,7 +1,7 @@
 import { useEffect, useRef, useState } from 'react';
 import SortOffersList from '../sort-offers-list/sort-offers-list';
 import { useAppSelector } from '../../hooks';
-import { selectCurrentSortOffers } from '../../store/offers-process/selectors';
+import { selectCurrentSortOffers } from '../../store/offers-process/offers-process.selectors';
 
 function PlacesSortForm(): JSX.Element{
   const sortSpanRef = useRef<HTMLElement>(null);
