+++ src/services/local-storage.ts
@@ -1,6 +1,6 @@
 import { SortOption } from '../const';
 import { store } from '../store';
-import { changeCity, changeSortOffers } from '../store/offers-process/offers-process';
+import { changeCity, changeSortOffers } from '../store/offers-process/offers-process.slice';
 import { getLocalStoredString } from '../utils/local-storage';
 
 const storedCurrentCity = getLocalStoredString('currentCity');
