+++ src/store/offers-process/offers-process.selectors.ts
