+++ src/store/offers-process/offers-process.slice.ts
